Move the GitUtils into a Git folder of the main project. Remove the GitUtils project.
