+++ GoogleCloudExtension/GoogleCloudExtension.sln
@@ -39,8 +39,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoogleAnalyticsUtilsTests",
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoogleCloudExtensionUnitTests", "GoogleCloudExtensionUnitTests\GoogleCloudExtensionUnitTests.csproj", "{0CD177FD-16EE-4DE7-B013-5C26348A4AB9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoogleCloudExtension.GitUtils", "GoogleCloudExtension.GitUtils\GoogleCloudExtension.GitUtils.csproj", "{66B55AAF-8237-45CF-8462-6359318795AB}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoogleCloudExtension.Utils.UnitTests", "GoogleCloudExtension.Utils.UnitTests\GoogleCloudExtension.Utils.UnitTests.csproj", "{8F5877B1-4E7C-4427-A37F-C037A5D08B64}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoogleCloudExtension.TemplateWizards", "GoogleCloudExtension.TemplateWizards\GoogleCloudExtension.TemplateWizards.csproj", "{43DF89C4-AF9A-4F61-AFD5-E167E28FC649}"
@@ -109,10 +107,6 @@ Global
 		{0CD177FD-16EE-4DE7-B013-5C26348A4AB9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0CD177FD-16EE-4DE7-B013-5C26348A4AB9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0CD177FD-16EE-4DE7-B013-5C26348A4AB9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{66B55AAF-8237-45CF-8462-6359318795AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{66B55AAF-8237-45CF-8462-6359318795AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{66B55AAF-8237-45CF-8462-6359318795AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{66B55AAF-8237-45CF-8462-6359318795AB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{8F5877B1-4E7C-4427-A37F-C037A5D08B64}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{8F5877B1-4E7C-4427-A37F-C037A5D08B64}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8F5877B1-4E7C-4427-A37F-C037A5D08B64}.Release|Any CPU.ActiveCfg = Release|Any CPU
