+++ GoogleCloudExtension/GoogleCloudExtension/GoogleCloudExtension.csproj
@@ -184,6 +184,11 @@
       <DependentUpon>CopyablePromptDialogWindowContent.xaml</DependentUpon>
     </Compile>
     <Compile Include="CopyablePrompt\CopyablePromptDialogWindowViewModel.cs" />
+    <Compile Include="Git\CsrGitUtils.cs" />
+    <Compile Include="Git\GitCommandException.cs" />
+    <Compile Include="Git\GitCommit.cs" />
+    <Compile Include="Git\GitRepository.cs" />
+    <Compile Include="Git\VSGitData.cs" />
     <Compile Include="Services\Configuration\AppEngineConfiguration.cs" />
     <Compile Include="Deployment\AppEngineFlexDeployment.cs" />
     <Compile Include="Deployment\GkeDeployment.cs" />
@@ -1152,10 +1157,6 @@
     <ProjectReference Include="..\GoogleCloudExtension.GCloud\GoogleCloudExtension.GCloud.csproj">
       <Project>{95efac7e-4f6e-46f0-bcc8-90ee1487e1e1}</Project>
       <Name>GoogleCloudExtension.GCloud</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\GoogleCloudExtension.GitUtils\GoogleCloudExtension.GitUtils.csproj">
-      <Project>{66b55aaf-8237-45cf-8462-6359318795ab}</Project>
-      <Name>GoogleCloudExtension.GitUtils</Name>
     </ProjectReference>
     <ProjectReference Include="..\GoogleCloudExtension.Interop\GoogleCloudExtension.Interop.csproj">
       <Project>{856B9F2E-441F-405A-9D5C-AF5DC5653902}</Project>
